requirements.txt
apache-airflow[celery,postgres,redis]==2.6.0
pandas==2.2.3
azure-identity
azure-storage-blob
